Firebase設定取得処理を共通化
