+++ app/api/fcm.ts
@@ -3,6 +3,7 @@ import { z } from "zod";
 import { zValidator } from "@hono/zod-validator";
 import authRequired from "./utils/auth.ts";
 import { getEnv } from "../../shared/config.ts";
+import { parseFirebaseClientConfig } from "../../shared/firebase_config.ts";
 import { registerToken, unregisterToken } from "./services/fcm.ts";
 
 const app = new Hono();
@@ -11,33 +12,7 @@ app.use("/fcm/*", authRequired);
 
 app.get("/fcm/config", (c) => {
   const env = getEnv(c);
-  let firebaseConfig: Record<string, string> | null = null;
-  if (env["FIREBASE_CLIENT_CONFIG"]) {
-    try {
-      firebaseConfig = JSON.parse(env["FIREBASE_CLIENT_CONFIG"]);
-    } catch {
-      firebaseConfig = null;
-    }
-  } else {
-    const keys = [
-      "FIREBASE_API_KEY",
-      "FIREBASE_AUTH_DOMAIN",
-      "FIREBASE_PROJECT_ID",
-      "FIREBASE_STORAGE_BUCKET",
-      "FIREBASE_MESSAGING_SENDER_ID",
-      "FIREBASE_APP_ID",
-    ];
-    if (keys.every((k) => env[k])) {
-      firebaseConfig = {
-        apiKey: env["FIREBASE_API_KEY"],
-        authDomain: env["FIREBASE_AUTH_DOMAIN"],
-        projectId: env["FIREBASE_PROJECT_ID"],
-        storageBucket: env["FIREBASE_STORAGE_BUCKET"],
-        messagingSenderId: env["FIREBASE_MESSAGING_SENDER_ID"],
-        appId: env["FIREBASE_APP_ID"],
-      };
-    }
-  }
+  const firebaseConfig = parseFirebaseClientConfig(env);
   if (!firebaseConfig) return c.json({});
   return c.json({
     firebase: firebaseConfig,
